Prefer safe code for i16 -> u8 conversion
**This Commit**
Uses `to_le_bytes` to safely convert the `i16` samples into pairs of
`u8`s to stream as audio.

**Why?**
Mostly because removing the `unsafe` code will be a little less scary
for most readers. I also wrote a quick benchmark and, because we save on
the `buffer.to_vec()` call, it seems to be something like 5-10% faster
so that's nice.

**Note**
We use `to_le_bytes` specifically for the PCM data. For more details see
[this comment][0].

[0]: https://github.com/deepgram-devs/bevy-deepgram/pull/2#discussion_r867226894